decouple new options interface from MetadataOptions
